Added db log files to ignore.
Renamed topics controller to topic.  Added admin controller but it does nothing right now.  Modified some views.  Created a ConfigBean to be loaded at app init.  Tried to use the ConfigBean to setup ORM and datasources but found that the bean was being created on every request which kind of defeats the purpose of sticking it in the variables scope.